refactor: extract organization http feature into ferriskey-api-organization
