+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["core", "api", "operator","client", "libs/maskass", "libs/ferriskey-security", "libs/ferriskey-domain", "libs/ferriskey-trident", "libs/ferriskey-abyss", "libs/ferriskey-aegis", "libs/ferriskey-compass", "libs/ferriskey-mail", "libs/ferriskey-migrate", "libs/ferriskey-organization", "libs/ferriskey-seawatch", "libs/ferriskey-password-policy", "libs/ferriskey-webhook", "libs/ferriskey-portal-theme", "libs/ferriskey-portal-layouts", "libs/ferriskey-api-core", "libs/ferriskey-api-abyss", "libs/ferriskey-api-seawatch", "libs/ferriskey-api-broker", "libs/ferriskey-api-compass", "libs/ferriskey-api-webhook", "libs/ferriskey-api-health", "libs/ferriskey-api-role", "libs/ferriskey-api-aegis", "libs/ferriskey-api-email-template", "libs/ferriskey-api-maintenance", "libs/ferriskey-api-portal-layouts", "libs/ferriskey-api-portal-theme", "libs/ferriskey-api-realm", "libs/ferriskey-api-client"]
+members = ["core", "api", "operator","client", "libs/maskass", "libs/ferriskey-security", "libs/ferriskey-domain", "libs/ferriskey-trident", "libs/ferriskey-abyss", "libs/ferriskey-aegis", "libs/ferriskey-compass", "libs/ferriskey-mail", "libs/ferriskey-migrate", "libs/ferriskey-organization", "libs/ferriskey-seawatch", "libs/ferriskey-password-policy", "libs/ferriskey-webhook", "libs/ferriskey-portal-theme", "libs/ferriskey-portal-layouts", "libs/ferriskey-api-core", "libs/ferriskey-api-abyss", "libs/ferriskey-api-seawatch", "libs/ferriskey-api-broker", "libs/ferriskey-api-compass", "libs/ferriskey-api-webhook", "libs/ferriskey-api-health", "libs/ferriskey-api-role", "libs/ferriskey-api-aegis", "libs/ferriskey-api-email-template", "libs/ferriskey-api-maintenance", "libs/ferriskey-api-portal-layouts", "libs/ferriskey-api-portal-theme", "libs/ferriskey-api-realm", "libs/ferriskey-api-client", "libs/ferriskey-api-organization"]
 resolver = "2"
 
 [workspace.package]
