+++ api/src/application/http/organization.rs
@@ -1,3 +1,1 @@
-pub mod handlers;
-pub mod router;
-mod validators;
+pub use ferriskey_api_organization::*;
