+++ libs/ferriskey-api-organization/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ferriskey-api-organization"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-api-core = { path = "../ferriskey-api-core" }
+ferriskey-core = { path = "../../core" }
+axum = { workspace = true }
+serde = { workspace = true }
+utoipa = { workspace = true }
+uuid = { workspace = true }
+validator = { workspace = true }
