+++ libs/ferriskey-api-organization/src/handlers/create_organization.rs
@@ -4,52 +4,46 @@ use axum::{
 };
 use ferriskey_core::domain::{
     authentication::value_objects::Identity,
-    organization::ports::{
-        Organization, OrganizationId, OrganizationService, UpdateOrganizationInput,
-    },
+    organization::ports::{CreateOrganizationInput, Organization, OrganizationService},
 };
-use uuid::Uuid;
 
-use crate::application::http::organization::validators::UpdateOrganizationValidator;
-use crate::application::http::server::api_entities::{
+use crate::validators::CreateOrganizationValidator;
+use ferriskey_api_core::api_entities::{
     api_error::{ApiError, ApiErrorResponse, ValidateJson},
     response::Response,
 };
-use crate::application::http::server::app_state::AppState;
+use ferriskey_api_core::app_state::AppState;
 
 #[utoipa::path(
-    put,
-    path = "/{organization_id}",
+    post,
+    path = "",
     tag = "organization",
-    summary = "Update organization details",
-    request_body = UpdateOrganizationValidator,
+    summary = "Create a new organization",
+    request_body = CreateOrganizationValidator,
     params(
         ("realm_name" = String, Path, description = "Realm name"),
-        ("organization_id" = Uuid, Path, description = "Organization ID"),
     ),
     responses(
-        (status = 200, description = "Organization updated successfully", body = Organization),
+        (status = 201, description = "Organization created successfully", body = Organization),
         (status = 400, description = "Invalid request data", body = ApiErrorResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
-        (status = 404, description = "Organization not found", body = ApiErrorResponse),
         (status = 422, description = "Validation error", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn update_organization(
-    Path((realm_name, organization_id)): Path<(String, Uuid)>,
+pub async fn create_organization(
+    Path(realm_name): Path<String>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-    ValidateJson(payload): ValidateJson<UpdateOrganizationValidator>,
+    ValidateJson(payload): ValidateJson<CreateOrganizationValidator>,
 ) -> Result<Response<Organization>, ApiError> {
     state
         .service
-        .update_organization(
+        .create_organization(
             identity,
-            UpdateOrganizationInput {
+            CreateOrganizationInput {
                 realm_name,
-                organization_id: OrganizationId::new(organization_id),
                 name: payload.name,
                 alias: payload.alias,
                 domain: payload.domain,
@@ -59,6 +53,6 @@ pub async fn update_organization(
             },
         )
         .await
-        .map(Response::Updated)
+        .map(Response::Created)
         .map_err(ApiError::from)
 }
