+++ libs/ferriskey-api-organization/src/handlers/list_attributes.rs
@@ -1,0 +1,60 @@
+use axum::{
+    Extension,
+    extract::{Path, State},
+};
+use ferriskey_core::domain::{
+    authentication::value_objects::Identity,
+    organization::ports::{
+        ListOrganizationAttributesInput, OrganizationAttribute, OrganizationId, OrganizationService,
+    },
+};
+use serde::Serialize;
+use utoipa::ToSchema;
+use uuid::Uuid;
+
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ApiErrorResponse},
+    response::Response,
+};
+use ferriskey_api_core::app_state::AppState;
+
+#[derive(Debug, Serialize, ToSchema)]
+pub struct ListOrganizationAttributesResponse {
+    pub data: Vec<OrganizationAttribute>,
+}
+
+#[utoipa::path(
+    get,
+    path = "/{organization_id}/attributes",
+    tag = "organization",
+    summary = "List organization attributes",
+    params(
+        ("realm_name" = String, Path, description = "Realm name"),
+        ("organization_id" = Uuid, Path, description = "Organization ID"),
+    ),
+    responses(
+        (status = 200, description = "Attributes retrieved successfully", body = ListOrganizationAttributesResponse),
+        (status = 401, description = "Unauthorized", body = ApiErrorResponse),
+        (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
+        (status = 404, description = "Organization not found", body = ApiErrorResponse),
+        (status = 500, description = "Internal server error", body = ApiErrorResponse),
+    ),
+)]
+pub async fn list_attributes(
+    Path((realm_name, organization_id)): Path<(String, Uuid)>,
+    State(state): State<AppState>,
+    Extension(identity): Extension<Identity>,
+) -> Result<Response<ListOrganizationAttributesResponse>, ApiError> {
+    state
+        .service
+        .list_attributes(
+            identity,
+            ListOrganizationAttributesInput {
+                realm_name,
+                organization_id: OrganizationId::new(organization_id),
+            },
+        )
+        .await
+        .map(|data| Response::OK(ListOrganizationAttributesResponse { data }))
+        .map_err(ApiError::from)
+}
