+++ libs/ferriskey-api-organization/src/handlers/list_members.rs
@@ -10,11 +10,11 @@ use ferriskey_core::domain::{
 };
 use uuid::Uuid;
 
-use crate::application::http::server::api_entities::{
+use ferriskey_api_core::api_entities::{
     api_error::{ApiError, ApiErrorResponse},
     response::Response,
 };
-use crate::application::http::server::app_state::AppState;
+use ferriskey_api_core::app_state::AppState;
 
 #[utoipa::path(
     get,
