+++ libs/ferriskey-api-organization/src/handlers/update_organization.rs
@@ -5,56 +5,60 @@ use axum::{
 use ferriskey_core::domain::{
     authentication::value_objects::Identity,
     organization::ports::{
-        AddOrganizationMemberInput, OrganizationId, OrganizationMember, OrganizationService,
+        Organization, OrganizationId, OrganizationService, UpdateOrganizationInput,
     },
 };
 use uuid::Uuid;
 
-use crate::application::http::organization::validators::AddMemberValidator;
-use crate::application::http::server::api_entities::{
+use crate::validators::UpdateOrganizationValidator;
+use ferriskey_api_core::api_entities::{
     api_error::{ApiError, ApiErrorResponse, ValidateJson},
     response::Response,
 };
-use crate::application::http::server::app_state::AppState;
+use ferriskey_api_core::app_state::AppState;
 
 #[utoipa::path(
-    post,
-    path = "/{organization_id}/members",
+    put,
+    path = "/{organization_id}",
     tag = "organization",
-    summary = "Add a member to an organization",
-    request_body = AddMemberValidator,
+    summary = "Update organization details",
+    request_body = UpdateOrganizationValidator,
     params(
         ("realm_name" = String, Path, description = "Realm name"),
         ("organization_id" = Uuid, Path, description = "Organization ID"),
     ),
     responses(
-        (status = 201, description = "Member added successfully", body = OrganizationMember),
+        (status = 200, description = "Organization updated successfully", body = Organization),
         (status = 400, description = "Invalid request data", body = ApiErrorResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
-        (status = 404, description = "Organization or user not found", body = ApiErrorResponse),
-        (status = 409, description = "User is already a member", body = ApiErrorResponse),
+        (status = 404, description = "Organization not found", body = ApiErrorResponse),
         (status = 422, description = "Validation error", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn add_member(
+pub async fn update_organization(
     Path((realm_name, organization_id)): Path<(String, Uuid)>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-    ValidateJson(payload): ValidateJson<AddMemberValidator>,
-) -> Result<Response<OrganizationMember>, ApiError> {
+    ValidateJson(payload): ValidateJson<UpdateOrganizationValidator>,
+) -> Result<Response<Organization>, ApiError> {
     state
         .service
-        .add_member(
+        .update_organization(
             identity,
-            AddOrganizationMemberInput {
+            UpdateOrganizationInput {
                 realm_name,
                 organization_id: OrganizationId::new(organization_id),
-                user_id: payload.user_id,
+                name: payload.name,
+                alias: payload.alias,
+                domain: payload.domain,
+                redirect_url: payload.redirect_url,
+                description: payload.description,
+                enabled: payload.enabled,
             },
         )
         .await
-        .map(Response::Created)
+        .map(Response::Updated)
         .map_err(ApiError::from)
 }
